✅ Update tests after refactoring source
